+++ ZeneFit/Common/Extension/Collections+.swift
@@ -1,0 +1,14 @@
+//
+//  Collections.swift
+//  ZeneFit
+//
+//  Created by iOS신상우 on 2023/10/15.
+//
+
+import Foundation
+
+extension Collection {
+    var isNotEmpty: Bool {
+        !self.isEmpty
+    }
+}
