+++ ZeneFit/Common/Extension/UIView+.swift
@@ -19,5 +19,9 @@ extension UIView {
         
         layer.insertSublayer(gradientLayer, at: 0)
     }
+    
+    func publisher<G>(for gestureRecognizer: G) -> UIGestureRecognizer.Publisher<G> where G: UIGestureRecognizer {
+        UIGestureRecognizer.Publisher(gestureRecognizer: gestureRecognizer, view: self)
+    }
 }
