+++ ZeneFit/Common/UI/InfoSelectionBottomSheet/SelectionCell.swift
@@ -20,6 +20,7 @@ final class SelectionCell: UITableViewCell {
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         self.selectionStyle = .none
+        self.backgroundColor = .white
         
         [titleLabel, checkImage].forEach {
             addSubview($0)
